+++ chaincode/src/main/java/de/upb/cs/uc4/chaincode/error/LedgerStateNotFoundError.java
@@ -1,0 +1,7 @@
+package de.upb.cs.uc4.chaincode.error;
+
+public class LedgerStateNotFoundError extends LedgerAccessError {
+    public LedgerStateNotFoundError(String jsonError) {
+        super(jsonError);
+    }
+}