+++ chaincode/src/main/java/de/upb/cs/uc4/chaincode/util/ContractUtil.java
@@ -1,10 +1,15 @@
 package de.upb.cs.uc4.chaincode.util;
 
+import com.google.gson.reflect.TypeToken;
 import de.upb.cs.uc4.chaincode.model.DetailedError;
 import de.upb.cs.uc4.chaincode.model.GenericError;
 import de.upb.cs.uc4.chaincode.model.InvalidParameter;
 import org.hyperledger.fabric.shim.ChaincodeStub;
+import org.hyperledger.fabric.shim.ledger.CompositeKey;
+import org.hyperledger.fabric.shim.ledger.KeyValue;
+import org.hyperledger.fabric.shim.ledger.QueryResultsIterator;
 
+import java.lang.reflect.Type;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -25,18 +30,19 @@ abstract public class ContractUtil {
 
     public abstract GenericError getConflictError();
 
-    protected GenericError getConflictError(String thing) {
+    protected GenericError getConflictError(String thing, String identifier) {
+        String article = "aeio".contains(Character.toString(thing.charAt(0)).toLowerCase()) ? "an" : "a";
         return new GenericError()
                 .type("HLConflict")
-                .title("There is already a " + thing + " for the given enrollmentId");
+                .title("There is already " + article + " " + thing + " for the given " + identifier);
     }
 
     public abstract GenericError getNotFoundError();
 
-    protected GenericError getNotFoundError(String thing) {
+    protected GenericError getNotFoundError(String thing, String identifier) {
         return new GenericError()
                 .type("HLNotFound")
-                .title("There is no " + thing + " for the given enrollmentId");
+                .title("There is no " + thing + " for the given " + identifier);
     }
 
     public GenericError getUnprocessableLedgerStateError() {
@@ -56,12 +62,19 @@ abstract public class ContractUtil {
     }
 
     public String putAndGetStringState(ChaincodeStub stub, String key, String value) {
-        stub.putStringState(keyPrefix + key,value);
+        String fullKey = stub.createCompositeKey(keyPrefix, key).toString();
+        stub.putStringState(fullKey,value);
         return value;
     }
 
     public String getStringState(ChaincodeStub stub, String key) {
-        return stub.getStringState(keyPrefix + key);
+        String fullKey = stub.createCompositeKey(keyPrefix, key).toString();
+        return stub.getStringState(fullKey);
+    }
+
+    public QueryResultsIterator<KeyValue> getAllRawStates(ChaincodeStub stub) {
+        CompositeKey key = stub.createCompositeKey(keyPrefix);
+        return stub.getStateByPartialCompositeKey(key);
     }
 
     public String getKeyPrefix() {