+++ chaincode/src/test/java/de/upb/cs/uc4/chaincode/ApprovalContractTest.java
@@ -3,7 +3,7 @@ package de.upb.cs.uc4.chaincode;
 
 import com.google.gson.reflect.TypeToken;
 import de.upb.cs.uc4.chaincode.model.JsonIOTest;
-import de.upb.cs.uc4.chaincode.util.CertificateContractUtil;
+import de.upb.cs.uc4.chaincode.util.ApprovalContractUtil;
 import de.upb.cs.uc4.chaincode.util.GsonWrapper;
 import de.upb.cs.uc4.chaincode.util.TestUtil;
 import org.hyperledger.fabric.contract.Context;
@@ -20,14 +20,14 @@ import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-public final class CertificateContractTest {
+public final class ApprovalContractTest {
 
-    private final CertificateContract contract = new CertificateContract();
-    private final CertificateContractUtil cUtil = new CertificateContractUtil();
+    private final ApprovalContract contract = new ApprovalContract();
+    private final ApprovalContractUtil cUtil = new ApprovalContractUtil();
 
     @TestFactory
     List<DynamicTest> createTests() {
-        String testConfigDir = "src/test/resources/test_configs/certificate_contract";
+        String testConfigDir = "src/test/resources/test_configs/approval_contract";
         File dir = new File(testConfigDir);
         File[] testConfigs = dir.listFiles();
 
@@ -52,34 +52,22 @@ public final class CertificateContractTest {
                 List<String> input = TestUtil.toStringList(test.getInput());
                 List<String> compare = TestUtil.toStringList(test.getCompare());
                 switch (test.getType()) {
-                    case "getCertificate":
+                    case "getApprovals":
                         tests.add(DynamicTest.dynamicTest(
                                 test.getName(),
-                                getCertificateTest(setup, input, compare)
+                                getApprovalsTest(setup, input, compare)
                         ));
                         break;
-                    case "addCertificate_SUCCESS":
+                    case "approveTransaction_SUCCESS":
                         tests.add(DynamicTest.dynamicTest(
                                 test.getName(),
-                                addCertificateSuccessTest(setup, input, compare)
+                                approveTransactionSuccessTest(setup, input, compare)
                         ));
                         break;
-                    case "addCertificate_FAILURE":
+                    case "approveTransaction_FAILURE":
                         tests.add(DynamicTest.dynamicTest(
                                 test.getName(),
-                                addCertificateFailureTest(setup, input, compare)
-                        ));
-                        break;
-                    case "updateCertificate_SUCCESS":
-                        tests.add(DynamicTest.dynamicTest(
-                                test.getName(),
-                                updateCertificateSuccessTest(setup, input, compare)
-                        ));
-                        break;
-                    case "updateCertificate_FAILURE":
-                        tests.add(DynamicTest.dynamicTest(
-                                test.getName(),
-                                updateCertificateFailureTest(setup, input, compare)
+                                approveTransactionFailureTest(setup, input, compare)
                         ));
                         break;
                     default:
@@ -90,7 +78,7 @@ public final class CertificateContractTest {
         return tests;
     }
 
-    private Executable getCertificateTest(
+    private Executable getApprovalsTest(
             List<String> setup,
             List<String> input,
             List<String> compare
@@ -98,12 +86,12 @@ public final class CertificateContractTest {
         return () -> {
             Context ctx = TestUtil.mockContext(setup, cUtil);
 
-            String certificate = contract.getCertificate(ctx, input.get(0));
-            assertThat(certificate).isEqualTo(compare.get(0));
+            String approvals = contract.getApprovals(ctx, contract(input), transaction(input), params(input));
+            assertThat(approvals).isEqualTo(compare.get(0));
         };
     }
 
-    private Executable addCertificateSuccessTest(
+    private Executable approveTransactionSuccessTest(
             List<String> setup,
             List<String> input,
             List<String> compare
@@ -111,14 +99,15 @@ public final class CertificateContractTest {
         return () -> {
             Context ctx = TestUtil.mockContext(setup, cUtil);
 
-            assertThat(contract.addCertificate(ctx, input.get(0), input.get(1)))
+            assertThat(contract.approveTransaction(ctx, contract(input), transaction(input), params(input)))
                     .isEqualTo(compare.get(0));
-            assertThat(cUtil.getStringState(ctx.getStub(), input.get(0)))
+            String key = cUtil.getDraftKey(contract(input), transaction(input), params(input));
+            assertThat(cUtil.getStringState(ctx.getStub(), key))
                     .isEqualTo(compare.get(0));
         };
     }
 
-    private Executable addCertificateFailureTest(
+    private Executable approveTransactionFailureTest(
             List<String> setup,
             List<String> input,
             List<String> compare
@@ -126,37 +115,23 @@ public final class CertificateContractTest {
         return () -> {
             Context ctx = TestUtil.mockContext(setup, cUtil);
 
-            String result = contract.addCertificate(ctx, input.get(0), input.get(1));
+            String result = contract.approveTransaction(ctx, contract(input), transaction(input), params(input));
             assertThat(result).isEqualTo(compare.get(0));
         };
     }
 
-    private Executable updateCertificateSuccessTest(
-            List<String> setup,
-            List<String> input,
-            List<String> compare
-    ) {
-        return () -> {
-            Context ctx = TestUtil.mockContext(setup, cUtil);
-
-            assertThat(contract.updateCertificate(ctx, input.get(0), input.get(1)))
-                    .isEqualTo(compare.get(0));
-            assertThat(cUtil.getStringState(ctx.getStub(), input.get(0)))
-                    .isEqualTo(compare.get(0));
-        };
-
+    private String contract(List<String> input) {
+        return input.get(0);
     }
 
-    private Executable updateCertificateFailureTest(
-            List<String> setup,
-            List<String> input,
-            List<String> compare
-    ) {
-        return () -> {
-            Context ctx = TestUtil.mockContext(setup, cUtil);
+    private String transaction(List<String> input) {
+        return input.get(1);
+    }
 
-            String result = contract.updateCertificate(ctx, input.get(0), input.get(1));
-            assertThat(result).isEqualTo(compare.get(0));
-        };
+    private String[] params(List<String> input) {
+        List<String> paramList = input.subList(2, input.size());
+        String[] params = new String[paramList.size()];
+        paramList.toArray(params);
+        return params;
     }
 }